+++ src/rovr/functions/icons.py
@@ -1,4 +1,4 @@
-from functools import cache
+from functools import lru_cache
 from os import path
 
 from rovr.variables.constants import config
@@ -13,7 +13,7 @@ from rovr.variables.maps import (
 )
 
 
-@cache
+@lru_cache(maxsize=1024)
 def get_icon_for_file(location: str) -> list:
     """
     Get the icon and color for a file based on its name or extension.
@@ -63,7 +63,7 @@ def get_icon_for_file(location: str) -> list:
     return ICONS["file"]["default"]
 
 
-@cache
+@lru_cache(maxsize=1024)
 def get_icon_for_folder(location: str) -> list:
     """Get the icon and color for a folder based on its name.
 
@@ -104,7 +104,7 @@ def get_icon_for_folder(location: str) -> list:
         return ICONS["folder"]["default"]
 
 
-@cache
+@lru_cache(maxsize=1024)
 def get_icon(outer_key: str, inner_key: str) -> list:
     """Get an icon from double keys.
     Args:
@@ -120,7 +120,7 @@ def get_icon(outer_key: str, inner_key: str) -> list:
         return ICONS[outer_key][inner_key]
 
 
-@cache
+@lru_cache(maxsize=1024)
 def get_toggle_button_icon(key: str) -> str:
     if not config["interface"]["nerd_font"]:
         return ASCII_TOGGLE_BUTTON_ICONS[key]