+++ src/rovr/functions/path.py
@@ -89,6 +89,11 @@ async def open_file(app: App, filepath: str) -> None:
     system = os_type.lower()
     # check if it is available first
     if not path.exists(filepath):
+        app.notify(
+            f"File not found: {filepath}",
+            title="Open File",
+            severity="error"
+        )
         return
 
     try:
@@ -208,7 +213,7 @@ async def get_cwd_object(
     return folders, files
 
 
-def file_is_type(file_path: str) -> str:
+def file_is_type(file_path: str) -> Literal["unknown", "symlink", "directory", "junction", "file"]:
     """Get a given path's type
     Args:
         file_path(str): The file path to check
@@ -284,7 +289,7 @@ def get_recursive_files(
         list: A list of dictionaries, with a "path" key and "relative_loc" key for files
         list: A list of path strings that were involved in the file list.
     """
-    if file_is_type(object_path) != "folder":
+    if file_is_type(object_path) != "directory":
         if with_folders:
             return [
                 {